+++ extensions/placeholder/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Placeholder Changelog
 
-## [Initial Version] - 2022-05-09
+## [Initial Version] - 2022-05-11
 
-- Easy to use, stylish placeholders from Lorem Picsum
-- Just add your desired image size (width & height), and you'll get a placeholder image.
+- Easy to use, stylish placeholders from [Lorem Picsum](https://picsum.photos)
+- Just add your desired image size (width & height), and you'll get a stylish placeholder.